+++ Nos/Assets/Localization/Localized.swift
@@ -14,7 +14,7 @@ import Foundation
 extension Localized {
     static var localizableTypes: [any Localizable.Type] {
         // TODO: Can we compute this using CaseIterable and recursion?
-        [Localized.self, Localized.Reply.self]
+        [Localized.self, Localized.Reply.self, Localized.ImagePicker.self]
     }
 }
 
@@ -106,7 +106,6 @@ enum Localized: String, Localizable, CaseIterable {
     case youNeedToEnterAPrivateKeyBeforePosting = "You need to enter a private key in Settings before you can publish posts."
     case relayDoesNotSupportNIP40 = "This relay doesn't support expiring messages. Please select another one."
     case anyRelaysSupportingNIP40 = "None of your relays support expiring messages. Please add one and retry."
-    case nostrBuildHelp = "Use nostr.build to post image links."
     case noteDisappearsIn = "Note disappears in" // Choices like 15 mins, 1 hour, etc. appear after this text.
     case attachMedia = "Attach Media"
     case expirationDate = "Expiration Date"
@@ -221,6 +220,7 @@ enum Localized: String, Localizable, CaseIterable {
     case eventSource = "Raw Event"
     case loading = "Loading..."
     case viewSource = "View Source"
+    case select = "Select"
 }
 
 // MARK: - Replies
@@ -239,4 +239,22 @@ extension Localized {
         case mentionedYou = "mentioned you:"
     }
 }
+
+// MARK: - ImagePicker
+
+extension Localized {
+    
+    enum ImagePicker: String, Localizable, CaseIterable {
+        case camera = "Camera"
+        case cameraNotAvailable = "Camera is not available on this device"
+        case openSettingsMessage = "You can allow camera permissions by opening the Settings app."
+        case permissionsRequired = "Permissions required for {{ title }}"
+        case photoLibrary = "Photo Library"
+        case selectFrom = "Select from Photo Library"
+        case takePhoto = "Take photo with Camera"
+        case errorUploadingFile = "Error uploading the file"
+        case errorUploadingFileMessage = "An error was encountered when uploading the file you provided. Please try again."
+    }
+}
+
 // swiftlint:enable line_length identifier_name