+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fixed the launch screen layout on iPad
+- Multiple consecutive newlines will be replaced by a single new line in note content.
+- Leading and trailing whitespace will no longer be rendered in note content.
+- Removed the screen to fill out profile information from onboarding and replaced it with a call to action in the sidebar. 
 - Added automatic file uploads to nostr.build.
 
 ## [0.1 (59)] - 2023-07-21Z